Add new set of test data for Mongo 3.2
Between Mongo 2.4 and 3.2, the format of mongoexport changed (among
other things, it puts less whitespace in its exported JSON), so that
tests run against Mongo 3.2 using the patches in the data/ folder (which
were originally created with Mongo 2.4) would result in excessively
large patch files. With the mongo-3.2-data folder, we now have a set of
baseline files for an empty project that can be used to run tests
against Mongo 3.2 without excessive whitespace in the patch files.