Add command-line options for min/max model version
There have been a couple of times when Fieldworks crashed with an
intermittent error related to NDesk.DBus getting a corrupted message. If
that crash happened while we was using the wizard on model 69 or 70, it
would be nice if we didn't have to reproduce our work creating the data
for model 68. These extra command-line options make that possible.